Nu kan du ta bort användare som har aktiva guider

+++ src/controller/Controller.java
@@ -16,7 +16,6 @@ public class Controller {
     private DbCon con;
     private GuiUtilities util;
     private AdminFrame adminFrame;
-
     private HomePageFrame homePageFrame;
     private MakeGuideGui SkapaGuideGui;
 
@@ -109,8 +108,14 @@ public class Controller {
      * @param username Användarnamn för raden man vill ta bort.
      */
     public void btnAdminDeleteUser(String username) {
-        con.deleteAUser(username);
+        if (con.checkIfUserHaveGuides(username)) {
+            if (util.showConfirmationDialog("User have still active guides \n Do you want to remove all guides to?") == 1) {
+                con.deleteGuideBasedOnUsername(username);
+                con.deleteAUser(username);
+            }
+        }
         adminFrame.updateUserList(con.getUsersAndEmail());
+        adminFrame.updateGuideList(con.getAllGuides());
     }
 
     /**
